# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Voom Car Rental API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  message: string;
}

export interface MessageResponse {
  message: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface RegisterRequest {
  fullName: string;
  username: string;
  phoneNumber: string;
  password: string;
}

export interface LoginRequest {
  username: string;
  password: string;
}

export interface UpdateProfileRequest {
  fullName?: string;
  profilePicture?: string;
  phoneNumber?: string;
}

export interface User {
  id: number;
  username: string;
  fullName?: string | null;
  phoneNumber?: string | null;
  profilePicture?: string | null;
  role?: string | null;
  isHost?: boolean | null;
  isVerified?: boolean | null;
  verificationStatus?: string | null;
  createdAt?: string | null;
}

export interface Car {
  id: number;
  hostId: number;
  make: string;
  model: string;
  year: number;
  type: string;
  dailyRate: number;
  currency: string;
  location: string;
  city?: string | null;
  country?: string | null;
  description?: string | null;
  imageUrl?: string | null;
  images?: string[] | null;
  color?: string | null;
  licensePlate?: string | null;
  rating?: number | null;
  ratingCount?: number | null;
  available?: boolean | null;
  features?: string[] | null;
  status?: string | null;
  transmission?: string | null;
  fuelType?: string | null;
  seats?: number | null;
  createdAt?: string | null;
}

export interface CreateCarRequest {
  make: string;
  model: string;
  year: number;
  type: string;
  dailyRate: number;
  currency: string;
  location: string;
  city?: string;
  country?: string;
  description?: string;
  imageUrl?: string;
  images?: string[];
  color?: string;
  licensePlate?: string;
  features?: string[];
  transmission?: string;
  fuelType?: string;
  seats?: number;
}

export interface UpdateCarRequest {
  make?: string;
  model?: string;
  year?: number;
  type?: string;
  dailyRate?: number;
  currency?: string;
  location?: string;
  description?: string;
  imageUrl?: string;
  available?: boolean;
  status?: string;
  features?: string[];
  transmission?: string;
  fuelType?: string;
  seats?: number;
}

export interface Booking {
  id: number;
  carId: number;
  userId: number;
  hostId: number;
  startDate: string;
  endDate: string;
  pickupLocation: string;
  dropoffLocation: string;
  totalAmount: number;
  platformFee?: number | null;
  hostPayout?: number | null;
  currency: string;
  paymentMethod?: string | null;
  status: string;
  createdAt?: string | null;
}

export type BookingWithCar = Booking & {
  car?: Car;
};

export interface CreateBookingRequest {
  carId: number;
  startDate: string;
  endDate: string;
  pickupLocation: string;
  dropoffLocation: string;
  paymentMethod?: string;
}

export interface Favorite {
  id: number;
  userId: number;
  carId: number;
  createdAt?: string | null;
}

export interface Message {
  id: number;
  senderId: number;
  receiverId: number;
  bookingId?: number | null;
  content: string;
  read?: boolean | null;
  createdAt?: string | null;
}

export interface MessageThread {
  userId: number;
  username: string;
  fullName?: string | null;
  profilePicture?: string | null;
  lastMessage: string;
  lastMessageAt?: string | null;
  unreadCount: number;
}

export interface SendMessageRequest {
  receiverId: number;
  content: string;
  bookingId?: number;
}

export interface Review {
  id: number;
  bookingId: number;
  reviewerId: number;
  revieweeId: number;
  carId: number;
  rating: number;
  text?: string | null;
  createdAt?: string | null;
}

export interface CreateReviewRequest {
  bookingId: number;
  revieweeId: number;
  carId: number;
  rating: number;
  text?: string;
}

export interface PaymentResponse {
  paymentId: number;
  amount: number;
  currency: string;
  status: string;
  message?: string | null;
}

export type UpdatePhoneBody = {
  phoneNumber: string;
};

export type UpdateProfilePictureBody = {
  profilePicture: string;
};

export type GetCarsParams = {
  category?: string;
  minPrice?: number;
  maxPrice?: number;
  available?: boolean;
  searchQuery?: string;
  sort?: string;
  limit?: number;
  offset?: number;
};

export type AddFavoriteBody = {
  carId: number;
};

export type CheckFavorite200 = {
  isFavorite: boolean;
};

export type CreateStripePaymentBody = {
  bookingId: number;
  idempotencyKey?: string;
};

export type InitiateMomoPaymentBody = {
  bookingId: number;
  phoneNumber: string;
  idempotencyKey?: string;
};