# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Voom Car Rental API
 * OpenAPI spec version: 0.1.0
 */

export interface Booking {
  id: number;
  carId: number;
  userId: number;
  hostId: number;
  startDate: Date;
  endDate: Date;
  pickupLocation: string;
  dropoffLocation: string;
  totalAmount: number;
  platformFee?: number | null;
  hostPayout?: number | null;
  currency: string;
  paymentMethod?: string | null;
  status: string;
  createdAt?: Date | null;
}