# lib/api-zod/src/generated/types/createBookingRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Voom Car Rental API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateBookingRequest {
  carId: number;
  startDate: Date;
  endDate: Date;
  pickupLocation: string;
  dropoffLocation: string;
  paymentMethod?: string;
}