# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Voom Car Rental API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  username: string;
  fullName?: string | null;
  phoneNumber?: string | null;
  profilePicture?: string | null;
  role?: string | null;
  isHost?: boolean | null;
  isVerified?: boolean | null;
  verificationStatus?: string | null;
  createdAt?: Date | null;
}